refactor/rename "Statistics" to "Dashboard" and move it to /views
